seprate routers and  services from main file
