+++ main.py
@@ -8,40 +8,28 @@ Clean architecture with feature-based organization:
 - Comprehensive error handling and logging
 """
 
-import json
 import logging
-import uuid
 from collections.abc import AsyncGenerator
 from contextlib import asynccontextmanager
-from datetime import UTC, datetime
 from pathlib import Path
 
 import uvicorn
-from fastapi import Depends, FastAPI, HTTPException, Request, WebSocket, WebSocketDisconnect
+from fastapi import  FastAPI, Request, WebSocket, WebSocketDisconnect
 from fastapi.middleware.cors import CORSMiddleware
 from fastapi.responses import HTMLResponse, JSONResponse
 from fastapi.staticfiles import StaticFiles
 from langfuse import observe
 from pydantic import ValidationError
-from sqlalchemy.ext.asyncio import AsyncSession
 
 from common.config.settings import settings
 from common.exceptions import AppError
-from common.schemas import ErrorResponse, HealthResponse, SuccessResponse
-from database.config.engine import get_database_session, sessionmanager
-from fitness_tracking.repositories.cricket_repository import (
-    CricketAnalyticsRepository,
-    CricketCoachingRepository,
-    CricketMatchRepository,
-    RestDayRepository,
-)
-from fitness_tracking.repositories.fitness_repository import FitnessRepository
-from voice_processing.schemas.conversation import ActivityType
-from voice_processing.schemas.processing import WebSocketMessage
+from common.schemas import ErrorResponse
+from database.config.engine import  sessionmanager
+
 from voice_processing.services.audio_storage import audio_storage
-from voice_processing.services.conversation_service import conversation_service
-from voice_processing.services.openai_service import openai_service
 from voice_processing.websocket.manager import connection_manager
+
+from router import Session_Management,info,audio,entries,analytics,dashboard,health,websocket
 
 # Configure logging using settings
 logging.basicConfig(
@@ -175,7 +163,7 @@ async def internal_server_error_handler(_request: Request, exc: Exception) -> JS
     )
 
 
-# Routes
+#  default Routes
 @app.get("/", response_class=HTMLResponse)
 async def root() -> HTMLResponse:
     """Serve the main web interface."""
@@ -200,899 +188,15 @@ async def root() -> HTMLResponse:
     )
 
 
-@app.get("/api")
-async def api_info() -> SuccessResponse:
-    """Return basic application info."""
-    return SuccessResponse(
-        message=f"{settings.app.title} API",
-        data={
-            "version": settings.app.version,
-            "environment": settings.app.environment,
-            "status": "running",
-            "endpoints": {
-                "voice_websocket": "/ws/voice/{session_id}",
-                "health_check": "/health",
-                "create_session": "/api/sessions",
-                "audio_stats": "/api/audio/stats",
-            },
-        },
-    )
 
-
-@app.get("/health")
-async def health_check() -> HealthResponse:
-    """Comprehensive health check endpoint."""
-    try:
-        # Get database statistics
-        db_stats = await sessionmanager.get_stats()
-
-        # Get WebSocket connection count
-        ws_connections = connection_manager.get_connection_count()
-
-        return HealthResponse(
-            status="healthy",
-            database=db_stats,
-            websocket_connections=ws_connections,
-            timestamp=f"{datetime.now(UTC).isoformat()}Z",
-        )
-
-    except Exception as e:
-        logger.exception("Health check failed")
-        raise HTTPException(
-            status_code=503,
-            detail="Service temporarily unavailable",
-        ) from e
-
-
-@app.post("/api/sessions")
-async def create_session() -> SuccessResponse:
-    """Create a new voice session."""
-    session_id = str(uuid.uuid4())
-    logger.info("Created new session: %s", session_id)
-
-    return SuccessResponse(
-        message="Session created successfully",
-        data={
-            "session_id": session_id,
-            "websocket_url": f"/ws/voice/{session_id}",
-            "status": "created",
-            "settings": {
-                "max_file_size_mb": settings.audio.max_file_size_mb,
-                "supported_formats": settings.audio.supported_formats,
-                "max_duration_seconds": settings.audio.max_duration_seconds,
-            },
-        },
-    )
-
-
-@app.get("/api/sessions/{session_id}")
-async def get_session_info(session_id: str) -> SuccessResponse:
-    """Get information about a specific session."""
-    connection_info = connection_manager.get_connection_info(session_id)
-
-    if not connection_info:
-        raise HTTPException(status_code=404, detail="Session not found")
-
-    return SuccessResponse(
-        message="Session info retrieved",
-        data=connection_info,
-    )
-
-
-@app.get("/api/audio/stats")
-async def get_audio_storage_stats() -> SuccessResponse:
-    """Get audio storage statistics."""
-    try:
-        stats = audio_storage.get_storage_stats()
-        return SuccessResponse(
-            message="Audio storage statistics",
-            data=stats,
-        )
-    except Exception as e:
-        logger.exception("Failed to get audio storage stats")
-        raise HTTPException(status_code=500, detail="Failed to retrieve audio stats") from e
-
-
-@app.get("/api/audio/sessions/{session_id}")
-async def get_session_audio_files(session_id: str) -> SuccessResponse:
-    """Get audio files for a specific session."""
-    try:
-        files = audio_storage.get_session_audio_files(session_id)
-        return SuccessResponse(
-            message="Session audio files",
-            data=files,
-        )
-    except Exception as e:
-        logger.exception("Failed to get session audio files")
-        raise HTTPException(status_code=500, detail="Failed to retrieve session audio files") from e
-
-
-@app.websocket("/ws/voice/{session_id}")
-@observe(capture_input=False, capture_output=False)
-async def voice_websocket_endpoint(websocket: WebSocket, session_id: str) -> None:
-    """
-    Modern WebSocket endpoint for real-time voice processing.
-
-    Uses the new settings-based configuration and refactored services.
-    """
-    try:
-        # Validate session ID - use constant for minimum length
-        min_session_id_length = 8
-        if not session_id or len(session_id) < min_session_id_length:
-            await websocket.close(code=1008, reason="Invalid session ID")
-            return
-
-        # Accept connection through connection manager
-        await connection_manager.connect(websocket, session_id)
-
-        # Send welcome message with settings-based configuration
-        welcome_message = WebSocketMessage(
-            type="connection_established",
-            session_id=session_id,
-            data={
-                "supported_entry_types": [
-                    "fitness",
-                    "cricket_coaching",
-                    "cricket_match",
-                    "rest_day",
-                ],
-                "max_message_size": settings.websocket.max_message_size,
-                "ping_interval": settings.websocket.ping_interval,
-                "audio_settings": {
-                    "max_file_size_mb": settings.audio.max_file_size_mb,
-                    "supported_formats": settings.audio.supported_formats,
-                    "max_duration_seconds": settings.audio.max_duration_seconds,
-                },
-                "recording_instructions": {
-                    "chunk_accumulation": True,
-                    "completion_signal": "Send 'recording_complete' message to process accumulated audio",
-                },
-            },
-        )
-        await connection_manager.send_message(welcome_message, session_id)
-
-        # Main message processing loop
-        while True:
-            try:
-                # Receive message with size limit
-                data = await websocket.receive()
-
-                if data.get("type") == "websocket.disconnect":
-                    break
-
-                # Handle different message types
-                if "text" in data:
-                    await handle_text_message(session_id, data["text"])
-                elif "bytes" in data:
-                    # Accumulate audio chunks instead of processing immediately
-                    await handle_audio_chunk(session_id, data["bytes"])
-
-            except WebSocketDisconnect:
-                logger.info("WebSocket disconnected for session: %s", session_id)
-                break
-            except Exception:
-                logger.exception("Error processing WebSocket message")
-                error_message = WebSocketMessage(
-                    type="error",
-                    session_id=session_id,
-                    error="message_processing_failed",
-                    message="Failed to process WebSocket message",
-                )
-                try:
-                    await connection_manager.send_message(error_message, session_id)
-                except Exception:
-                    break
-
-    except Exception:
-        logger.exception("WebSocket connection error for session %s", session_id)
-    finally:
-        connection_manager.disconnect(session_id)
-
-
-async def handle_text_message(session_id: str, message: str) -> None:
-    """Handle text messages from WebSocket clients."""
-    try:
-        message_data = json.loads(message)
-        message_type = message_data.get("type")
-
-        if message_type == "voice_data_meta":
-            # Store metadata for this session
-            entry_type = message_data.get("entry_type")
-            user_id = message_data.get("user_id", "demo_user")
-
-            # Validate entry type
-            valid_entry_types = ["fitness", "cricket_coaching", "cricket_match", "rest_day"]
-            if entry_type not in valid_entry_types:
-                error_message = WebSocketMessage(
-                    type="error",
-                    session_id=session_id,
-                    error="invalid_entry_type",
-                    message=f"Invalid entry type. Must be one of: {valid_entry_types}",
-                )
-                await connection_manager.send_message(error_message, session_id)
-                return
-
-            # Store session metadata
-            connection_manager.set_session_metadata(
-                session_id,
-                {
-                    "entry_type": entry_type,
-                    "user_id": user_id,
-                },
-            )
-
-            # Acknowledge metadata received
-            response = WebSocketMessage(
-                type="voice_meta_received",
-                session_id=session_id,
-                data={
-                    "entry_type": entry_type,
-                    "user_id": user_id,
-                    "ready_for_audio": True,
-                },
-            )
-            await connection_manager.send_message(response, session_id)
-
-        elif message_type == "recording_complete":
-            # Process the accumulated audio when recording is complete
-            logger.info("Received recording_complete signal for session %s", session_id)
-            await handle_complete_audio_processing(session_id)
-
-        else:
-            logger.warning("Unknown text message type: %s", message_type)
-
-    except json.JSONDecodeError:
-        logger.exception("Invalid JSON in text message")
-        error_message = WebSocketMessage(
-            type="error",
-            session_id=session_id,
-            error="invalid_json",
-            message="Message must be valid JSON",
-        )
-        await connection_manager.send_message(error_message, session_id)
-
-
-@observe(capture_input=True, capture_output=False)
-async def handle_audio_chunk(session_id: str, audio_chunk: bytes) -> None:
-    """Accumulate audio chunks for later processing when recording is complete."""
-    try:
-        # Validate audio chunk
-        if len(audio_chunk) == 0:
-            logger.debug("Received empty audio chunk for session %s", session_id)
-            return
-
-        # Check maximum individual chunk size to prevent abuse
-        max_chunk_size = 2 * 1024 * 1024  # 2MB per chunk
-        if len(audio_chunk) > max_chunk_size:
-            error_message = WebSocketMessage(
-                type="error",
-                session_id=session_id,
-                error="audio_chunk_too_large",
-                message=f"Audio chunk exceeds maximum size: {len(audio_chunk)} bytes",
-            )
-            await connection_manager.send_message(error_message, session_id)
-            return
-
-        # Accumulate the audio chunk
-        connection_manager.accumulate_audio_chunk(session_id, audio_chunk)
-        total_size = len(connection_manager.get_accumulated_audio(session_id))
-
-        # Check total accumulated size
-        max_total_size = settings.audio.max_file_size_mb * 1024 * 1024
-        if total_size > max_total_size:
-            error_message = WebSocketMessage(
-                type="error",
-                session_id=session_id,
-                error="total_audio_too_large",
-                message=f"Total accumulated audio exceeds maximum size: {total_size} bytes",
-            )
-            await connection_manager.send_message(error_message, session_id)
-            # Clear the buffer to prevent further accumulation
-            connection_manager.clear_audio_buffer(session_id)
-            return
-
-        # Send acknowledgment for chunk received
-        chunk_ack = WebSocketMessage(
-            type="audio_chunk_received",
-            session_id=session_id,
-            data={
-                "chunk_size": len(audio_chunk),
-                "total_accumulated": total_size,
-                "status": "accumulated",
-            },
-        )
-        await connection_manager.send_message(chunk_ack, session_id)
-
-        logger.debug(
-            "Accumulated audio chunk for session %s: +%d bytes (total: %d bytes)",
-            session_id,
-            len(audio_chunk),
-            total_size,
-        )
-
-    except Exception as e:
-        logger.exception("Audio chunk processing failed for session %s", session_id)
-        error_message = WebSocketMessage(
-            type="error",
-            session_id=session_id,
-            error="audio_chunk_processing_failed",
-            message=str(e),
-        )
-        await connection_manager.send_message(error_message, session_id)
-
-
-@observe(capture_input=True, capture_output=False)
-async def handle_complete_audio_processing(session_id: str) -> None:
-    """Process complete accumulated audio with multi-turn conversation support."""
-    try:
-        start_time = datetime.now(UTC)
-
-        # Get accumulated audio data
-        voice_data = connection_manager.get_accumulated_audio(session_id)
-
-        if len(voice_data) == 0:
-            error_message = WebSocketMessage(
-                type="error",
-                session_id=session_id,
-                error="no_audio_data",
-                message="No audio data accumulated for processing",
-            )
-            await connection_manager.send_message(error_message, session_id)
-            return
-
-        # Check minimum audio size for meaningful content
-        min_audio_size = 1000  # At least 1KB for meaningful audio
-        if len(voice_data) < min_audio_size:
-            error_message = WebSocketMessage(
-                type="error",
-                session_id=session_id,
-                error="insufficient_audio_data",
-                message=f"Audio data too small: {len(voice_data)} bytes. Please record for longer.",
-            )
-            await connection_manager.send_message(error_message, session_id)
-            return
-
-        # Get session metadata (entry_type and user_id)
-        session_metadata = connection_manager.get_session_metadata(session_id)
-        if not session_metadata:
-            error_message = WebSocketMessage(
-                type="error",
-                session_id=session_id,
-                error="no_session_metadata",
-                message="Please send voice_data_meta message first with entry_type",
-            )
-            await connection_manager.send_message(error_message, session_id)
-            return
-
-        entry_type = session_metadata.get("entry_type")
-        user_id = session_metadata.get("user_id", "demo_user")
-
-        # Send processing start acknowledgment
-        processing_message = WebSocketMessage(
-            type="audio_processing_started",
-            session_id=session_id,
-            data={
-                "status": "processing_complete_audio",
-                "total_audio_size": len(voice_data),
-                "entry_type": entry_type,
-                "user_id": user_id,
-            },
-        )
-        await connection_manager.send_message(processing_message, session_id)
-
-        # 1. Save complete audio using refactored audio storage
-        logger.info(
-            "Processing complete audio for session %s: %s bytes, type: %s",
-            session_id,
-            len(voice_data),
-            entry_type,
-        )
-
-        audio_save_result = await audio_storage.save_raw_audio(
-            session_id=session_id,
-            audio_data=voice_data,
-            audio_format="webm",
-        )
-
-        # 2. Transcribe complete audio using improved OpenAI service
-        transcription_result = await openai_service.transcribe_audio(voice_data)
-
-        # Send transcript to user for confirmation
-        transcript_message = WebSocketMessage(
-            type="transcript_ready",
-            session_id=session_id,
-            data={
-                "transcript": transcription_result.text,
-                "confidence": transcription_result.confidence,
-                "language": transcription_result.language,
-                "entry_type": entry_type,
-            },
-        )
-        await connection_manager.send_message(transcript_message, session_id)
-
-        # 3. **NEW**: Use conversation service for multi-turn processing
-        # Map entry_type to ActivityType
-        activity_type_mapping = {
-            "fitness": ActivityType.FITNESS,
-            "cricket_coaching": ActivityType.CRICKET_COACHING,
-            "cricket_match": ActivityType.CRICKET_MATCH,
-            "rest_day": ActivityType.REST_DAY,
-        }
-        activity_type = activity_type_mapping.get(entry_type)
-
-        if not activity_type:
-            error_message = WebSocketMessage(
-                type="error",
-                session_id=session_id,
-                error="invalid_activity_type",
-                message=f"Unsupported activity type: {entry_type}",
-            )
-            await connection_manager.send_message(error_message, session_id)
-            return
-
-        # Start or continue conversation
-        conversation_context = conversation_service.get_conversation(session_id)
-        if not conversation_context:
-            # Start new conversation
-            conversation_context = conversation_service.start_conversation(
-                session_id=session_id,
-                user_id=user_id,
-                activity_type=activity_type,
-            )
-
-            # Send conversation started message
-            conversation_started_message = WebSocketMessage(
-                type="conversation_started",
-                session_id=session_id,
-                data={
-                    "activity_type": activity_type.value,
-                    "message": "I'll help you log your activity. Let's start with what you've told me and I'll ask follow-up questions to get all the details.",
-                },
-            )
-            await connection_manager.send_message(conversation_started_message, session_id)
-
-        # Process user input and get conversation analysis
-        conversation_analysis = await conversation_service.process_user_input(
-            session_id=session_id,
-            user_input=transcription_result.text,
-            transcript_confidence=transcription_result.confidence,
-        )
-
-        # 4. Determine next step based on conversation analysis
-        if conversation_analysis.should_continue and conversation_analysis.next_question:
-            # Ask follow-up question
-            follow_up_message = WebSocketMessage(
-                type="follow_up_question",
-                session_id=session_id,
-                data={
-                    "question": conversation_analysis.next_question.question,
-                    "field_target": conversation_analysis.next_question.field_target,
-                    "question_type": conversation_analysis.next_question.question_type,
-                    "priority": conversation_analysis.next_question.priority,
-                    "turn_number": conversation_context.turn_count,
-                    "completeness_score": conversation_analysis.data_completeness.completeness_score,
-                    "collected_data": conversation_context.collected_data,
-                    "missing_fields": conversation_analysis.data_completeness.missing_fields,
-                    "instructions": "Please answer the question and then tap the microphone to record your response.",
-                },
-            )
-            await connection_manager.send_message(follow_up_message, session_id)
-
-            logger.info(
-                "Sent follow-up question for session %s: %s (targeting field: %s)",
-                session_id,
-                conversation_analysis.next_question.question,
-                conversation_analysis.next_question.field_target,
-            )
-
-        elif conversation_analysis.can_generate_final_output:
-            # Complete conversation and save to database
-            conversation_result = conversation_service.complete_conversation(session_id)
-
-            # Save final structured data to database
-            structured_data = conversation_result.final_data
-            saved_entry = None
-            processing_duration = (datetime.now(UTC) - start_time).total_seconds()
-
-            # Get database session for saving
-            async with sessionmanager.get_session() as db_session:
-                from fitness_tracking.repositories.cricket_repository import (
-                    CricketCoachingRepository,
-                    CricketMatchRepository,
-                    RestDayRepository,
-                )
-                from fitness_tracking.repositories.fitness_repository import FitnessRepository
-
-                if entry_type == "fitness":
-                    fitness_repo = FitnessRepository(db_session)
-                    saved_entry = await fitness_repo.create_from_voice_data(
-                        session_id=session_id,
-                        user_id=user_id,
-                        voice_data=structured_data,
-                        transcript=transcription_result.text,
-                        confidence_score=transcription_result.confidence,
-                        processing_duration=processing_duration,
-                    )
-
-                elif entry_type == "cricket_coaching":
-                    cricket_repo = CricketCoachingRepository(db_session)
-                    saved_entry = await cricket_repo.create_from_voice_data(
-                        session_id=session_id,
-                        user_id=user_id,
-                        voice_data=structured_data,
-                        transcript=transcription_result.text,
-                        confidence_score=transcription_result.confidence,
-                        processing_duration=processing_duration,
-                    )
-
-                elif entry_type == "cricket_match":
-                    match_repo = CricketMatchRepository(db_session)
-                    saved_entry = await match_repo.create_from_voice_data(
-                        session_id=session_id,
-                        user_id=user_id,
-                        voice_data=structured_data,
-                        transcript=transcription_result.text,
-                        confidence_score=transcription_result.confidence,
-                        processing_duration=processing_duration,
-                    )
-
-                elif entry_type == "rest_day":
-                    rest_repo = RestDayRepository(db_session)
-                    saved_entry = await rest_repo.create_from_voice_data(
-                        session_id=session_id,
-                        user_id=user_id,
-                        voice_data=structured_data,
-                        transcript=transcription_result.text,
-                        confidence_score=transcription_result.confidence,
-                        processing_duration=processing_duration,
-                    )
-
-            # Send conversation completed message
-            result_message = WebSocketMessage(
-                type="conversation_completed",
-                session_id=session_id,
-                data={
-                    "status": "success",
-                    "entry_type": entry_type,
-                    "structured_data": structured_data,
-                    "saved_entry_id": saved_entry.id if saved_entry else None,
-                    "audio_file": audio_save_result.get("filename"),
-                    "total_turns": conversation_result.total_turns,
-                    "data_quality_score": conversation_result.data_quality_score,
-                    "conversation_efficiency": conversation_result.conversation_efficiency,
-                    "processing_time": processing_duration,
-                    "database_saved": saved_entry is not None,
-                    "user_id": user_id,
-                    "message": f"Great! I've saved your {entry_type.replace('_', ' ')} entry with {len(structured_data)} data points collected over {conversation_result.total_turns} turns.",
-                },
-            )
-            await connection_manager.send_message(result_message, session_id)
-
-            logger.info(
-                "✅ Conversation completed for session %s: %d turns, %.2f quality, %.2f efficiency",
-                session_id,
-                conversation_result.total_turns,
-                conversation_result.data_quality_score,
-                conversation_result.conversation_efficiency,
-            )
-
-        else:
-            # Error case - something went wrong
-            error_message = WebSocketMessage(
-                type="error",
-                session_id=session_id,
-                error="conversation_analysis_failed",
-                message=f"Conversation analysis failed: {conversation_analysis.reasoning}",
-            )
-            await connection_manager.send_message(error_message, session_id)
-            conversation_service.cleanup_session(session_id)
-
-        # Clear the audio buffer after processing
-        connection_manager.clear_audio_buffer(session_id)
-
-    except Exception as e:
-        logger.exception("Complete audio processing failed for session %s", session_id)
-        error_message = WebSocketMessage(
-            type="error",
-            session_id=session_id,
-            error="voice_processing_failed",
-            message=str(e),
-        )
-        await connection_manager.send_message(error_message, session_id)
-        # Clean up on error
-        connection_manager.clear_audio_buffer(session_id)
-        conversation_service.cleanup_session(session_id)
-
-
-@app.get("/api/entries/fitness")
-async def get_fitness_entries(
-    limit: int = 10,
-    days_back: int = 30,
-    user_id: str = "demo_user",  # TODO: Get from authentication
-    session: AsyncSession = Depends(get_database_session),
-) -> SuccessResponse:
-    """Get recent fitness entries for a user."""
-    try:
-        fitness_repo = FitnessRepository(session)
-        entries = await fitness_repo.get_recent_entries(
-            user_id=user_id,
-            limit=limit,
-            days_back=days_back,
-        )
-
-        return SuccessResponse(
-            message="Fitness entries retrieved successfully",
-            data={
-                "entries": [entry.model_dump() for entry in entries],
-                "total_count": len(entries),
-                "user_id": user_id,
-                "days_back": days_back,
-            },
-        )
-    except Exception as e:
-        logger.exception("Failed to get fitness entries")
-        raise HTTPException(
-            status_code=500,
-            detail=f"Failed to retrieve fitness entries: {e}",
-        ) from e
-
-
-@app.get("/api/entries/cricket/coaching")
-async def get_cricket_coaching_entries(
-    limit: int = 10,
-    user_id: str = "demo_user",  # TODO: Get from authentication
-    session: AsyncSession = Depends(get_database_session),
-) -> SuccessResponse:
-    """Get cricket coaching session entries for a user."""
-    try:
-        cricket_repo = CricketCoachingRepository(session)
-        entries = await cricket_repo.read_multi(limit=limit)
-
-        return SuccessResponse(
-            message="Cricket coaching entries retrieved successfully",
-            data={
-                "entries": [entry.model_dump() for entry in entries],
-                "total_count": len(entries),
-                "user_id": user_id,
-            },
-        )
-    except Exception as e:
-        logger.exception("Failed to get cricket coaching entries")
-        raise HTTPException(
-            status_code=500,
-            detail=f"Failed to retrieve cricket coaching entries: {e}",
-        ) from e
-
-
-@app.get("/api/entries/cricket/matches")
-async def get_cricket_match_entries(
-    limit: int = 10,
-    user_id: str = "demo_user",  # TODO: Get from authentication
-    session: AsyncSession = Depends(get_database_session),
-) -> SuccessResponse:
-    """Get cricket match performance entries for a user."""
-    try:
-        match_repo = CricketMatchRepository(session)
-        entries = await match_repo.read_multi(limit=limit)
-
-        return SuccessResponse(
-            message="Cricket match entries retrieved successfully",
-            data={
-                "entries": [entry.model_dump() for entry in entries],
-                "total_count": len(entries),
-                "user_id": user_id,
-            },
-        )
-    except Exception as e:
-        logger.exception("Failed to get cricket match entries")
-        raise HTTPException(
-            status_code=500,
-            detail=f"Failed to retrieve cricket match entries: {e}",
-        ) from e
-
-
-@app.get("/api/entries/rest-days")
-async def get_rest_day_entries(
-    limit: int = 10,
-    user_id: str = "demo_user",  # TODO: Get from authentication
-    session: AsyncSession = Depends(get_database_session),
-) -> SuccessResponse:
-    """Get rest day entries for a user."""
-    try:
-        rest_repo = RestDayRepository(session)
-        entries = await rest_repo.read_multi(limit=limit)
-
-        return SuccessResponse(
-            message="Rest day entries retrieved successfully",
-            data={
-                "entries": [entry.model_dump() for entry in entries],
-                "total_count": len(entries),
-                "user_id": user_id,
-            },
-        )
-    except Exception as e:
-        logger.exception("Failed to get rest day entries")
-        raise HTTPException(
-            status_code=500,
-            detail=f"Failed to retrieve rest day entries: {e}",
-        ) from e
-
-
-@app.get("/api/analytics/fitness")
-async def get_fitness_analytics(
-    days_back: int = 30,
-    user_id: str = "demo_user",  # TODO: Get from authentication
-    session: AsyncSession = Depends(get_database_session),
-) -> SuccessResponse:
-    """Get comprehensive fitness analytics for a user."""
-    try:
-        fitness_repo = FitnessRepository(session)
-        analytics = await fitness_repo.get_fitness_analytics(
-            user_id=user_id,
-            days_back=days_back,
-        )
-
-        return SuccessResponse(
-            message="Fitness analytics generated successfully",
-            data={
-                "analytics": analytics.model_dump(),
-                "user_id": user_id,
-                "period_days": days_back,
-                "generated_at": datetime.now(UTC).isoformat(),
-            },
-        )
-    except Exception as e:
-        logger.exception("Failed to generate fitness analytics")
-        raise HTTPException(
-            status_code=500,
-            detail=f"Failed to generate fitness analytics: {e}",
-        ) from e
-
-
-@app.get("/api/analytics/cricket")
-async def get_cricket_analytics(
-    days_back: int = 30,
-    user_id: str = "demo_user",  # TODO: Get from authentication
-    session: AsyncSession = Depends(get_database_session),
-) -> SuccessResponse:
-    """Get comprehensive cricket analytics for a user."""
-    try:
-        cricket_analytics_repo = CricketAnalyticsRepository(session)
-        analytics = await cricket_analytics_repo.get_cricket_analytics(
-            user_id=user_id,
-            days_back=days_back,
-        )
-
-        return SuccessResponse(
-            message="Cricket analytics generated successfully",
-            data={
-                "analytics": analytics.model_dump(),
-                "user_id": user_id,
-                "period_days": days_back,
-                "generated_at": datetime.now(UTC).isoformat(),
-            },
-        )
-    except Exception as e:
-        logger.exception("Failed to generate cricket analytics")
-        raise HTTPException(
-            status_code=500,
-            detail=f"Failed to generate cricket analytics: {e}",
-        ) from e
-
-
-@app.get("/api/analytics/combined")
-async def get_combined_analytics(
-    days_back: int = 30,
-    user_id: str = "demo_user",  # TODO: Get from authentication
-    session: AsyncSession = Depends(get_database_session),
-) -> SuccessResponse:
-    """Get combined fitness and cricket analytics with correlations."""
-    try:
-        # Get both analytics
-        fitness_repo = FitnessRepository(session)
-        cricket_analytics_repo = CricketAnalyticsRepository(session)
-
-        fitness_analytics = await fitness_repo.get_fitness_analytics(user_id, days_back)
-        cricket_analytics = await cricket_analytics_repo.get_cricket_analytics(user_id, days_back)
-
-        # Simple correlation analysis (could be enhanced)
-        correlations = {}
-        if fitness_analytics.weekly_frequency > 0 and cricket_analytics.average_self_assessment > 0:
-            correlations["fitness_frequency_vs_cricket_confidence"] = min(
-                fitness_analytics.weekly_frequency
-                / 7
-                * cricket_analytics.average_self_assessment
-                / 10,
-                1.0,
-            )
-
-        # Generate combined recommendations
-        combined_recommendations = []
-        combined_recommendations.extend(fitness_analytics.recommendations[:2])
-        combined_recommendations.extend(cricket_analytics.recommendations[:2])
-
-        if fitness_analytics.weekly_frequency < 3:
-            combined_recommendations.append(
-                "Increase fitness frequency to improve cricket performance correlation",
-            )
-
-        return SuccessResponse(
-            message="Combined analytics generated successfully",
-            data={
-                "fitness_analytics": fitness_analytics.model_dump(),
-                "cricket_analytics": cricket_analytics.model_dump(),
-                "correlations": correlations,
-                "combined_recommendations": combined_recommendations,
-                "user_id": user_id,
-                "period_days": days_back,
-                "generated_at": datetime.now(UTC).isoformat(),
-            },
-        )
-    except Exception as e:
-        logger.exception("Failed to generate combined analytics")
-        raise HTTPException(
-            status_code=500,
-            detail=f"Failed to generate combined analytics: {e}",
-        ) from e
-
-
-@app.get("/api/dashboard")
-async def get_user_dashboard(
-    user_id: str = "demo_user",  # TODO: Get from authentication
-    session: AsyncSession = Depends(get_database_session),
-) -> SuccessResponse:
-    """Get comprehensive dashboard data for a user."""
-    try:
-        # Get recent entries from all types
-        fitness_repo = FitnessRepository(session)
-        cricket_coaching_repo = CricketCoachingRepository(session)
-        cricket_match_repo = CricketMatchRepository(session)
-        rest_repo = RestDayRepository(session)
-
-        # Get recent data (last 7 days)
-        recent_fitness = await fitness_repo.get_recent_entries(user_id, limit=5, days_back=7)
-        recent_coaching = await cricket_coaching_repo.read_multi(limit=3)
-        recent_matches = await cricket_match_repo.read_multi(limit=3)
-        recent_rest = await rest_repo.read_multi(limit=3)
-
-        # Get quick analytics
-        fitness_analytics = await fitness_repo.get_fitness_analytics(user_id, days_back=7)
-
-        # Activity summary for this week
-        activity_summary = {
-            "fitness_sessions": len(recent_fitness),
-            "cricket_coaching_sessions": len(recent_coaching),
-            "matches_played": len(recent_matches),
-            "rest_days": len(recent_rest),
-            "average_energy_level": fitness_analytics.average_energy_level,
-            "weekly_frequency": fitness_analytics.weekly_frequency,
-        }
-
-        return SuccessResponse(
-            message="Dashboard data retrieved successfully",
-            data={
-                "user_id": user_id,
-                "activity_summary": activity_summary,
-                "recent_entries": {
-                    "fitness": [entry.model_dump() for entry in recent_fitness],
-                    "cricket_coaching": [entry.model_dump() for entry in recent_coaching],
-                    "cricket_matches": [entry.model_dump() for entry in recent_matches],
-                    "rest_days": [entry.model_dump() for entry in recent_rest],
-                },
-                "quick_insights": {
-                    "most_common_fitness_activity": fitness_analytics.most_common_activity,
-                    "fitness_improvement_trends": fitness_analytics.improvement_trends,
-                    "recommendations": fitness_analytics.recommendations[:3],
-                },
-                "generated_at": datetime.now(UTC).isoformat(),
-            },
-        )
-    except Exception as e:
-        logger.exception("Failed to get dashboard data")
-        raise HTTPException(
-            status_code=500,
-            detail=f"Failed to retrieve dashboard data: {e}",
-        ) from e
+app.include_router(info.router)
+app.include_router(health.router)
+app.include_router(Session_Management.router)
+app.include_router(audio.router)
+app.include_router(websocket.router)
+app.include_router(entries.router)
+app.include_router(analytics.router)
+app.include_router(dashboard.router)
 
 
 if __name__ == "__main__":
